Bug fixes and requests
* Moved to CRandom to Common lib
* Initialized CRandom with a time seed in CN64System constructor
* Added CRandom::state getters and setters to support saved states
