+++ Source/Common/Common.vcxproj
@@ -43,6 +43,7 @@
     <ClCompile Include="MemTest.cpp" />
     <ClCompile Include="path.cpp" />
     <ClCompile Include="Platform.cpp" />
+    <ClCompile Include="Random.cpp" />
     <ClCompile Include="stdafx.cpp">
       <PrecompiledHeader>Create</PrecompiledHeader>
     </ClCompile>
@@ -64,6 +65,7 @@
     <ClInclude Include="MemTest.h" />
     <ClInclude Include="path.h" />
     <ClInclude Include="Platform.h" />
+    <ClInclude Include="Random.h" />
     <ClInclude Include="SmartPointer.h" />
     <ClInclude Include="stdafx.h" />
     <ClInclude Include="StdString.h" />
