+++ Source/Project64-core/N64System/SystemGlobals.cpp
@@ -32,5 +32,6 @@ uint32_t      * g_TLBStoreAddress = NULL;
 CDebugger     * g_Debugger = NULL;
 uint8_t      ** g_RecompPos = NULL;
 CMempak       * g_Mempak = NULL;
+CRandom       * g_Random = NULL;
 
 int * g_NextTimer;
